Back-End/requirements.txt
fastapi>=0.115.0,<0.136.0
uvicorn[standard]>=0.32.0,<0.36.0
sqlalchemy>=2.0.0,<2.1.0
httpx>=0.27.0,<0.29.0
openai>=1.59.0,<3.0.0
pydantic-settings>=2.7.0,<3.0.0
pytest>=8.3.0,<10.0.0
pytest-asyncio>=0.25.0,<2.0.0
anyio>=4.8.0,<5.0.0
PyJWT>=2.8.0,<3.0.0
passlib[bcrypt]>=1.7.4,<1.8.0
email-validator>=2.0.0,<3.0.0
